requirements.txt
numpy
opencv-python
Pillow
pdfplumber
matplotlib
tqdm